Tidy RSpec output; move helper methods to RSpec shared contexts
