+++ corpus-client/spec/client-emails-notes_spec.rb
@@ -1,46 +1,9 @@
 # n.b. see spec_helper for imports and constants
 
 feature "client can view the contents of the selected email" do
-  # Run/kill the elm-reactor
-  around(:example) do |example|
-    puts "running elm reactor on port #{ELM_REACTOR_PORT}"
-    tmp_elm_out = Tempfile.new("rspec-elm-out")
-    tmp_elm_err = Tempfile.new("rspec-elm-err")
-    elm_server_pid = Process.spawn(
-      "elm reactor --port=#{ELM_REACTOR_PORT}",
-      out: tmp_elm_out.path,
-      err: tmp_elm_err.path
-    )
-
-    Capybara.app_host = "http://localhost:#{ELM_REACTOR_PORT}"
-
-    example.run
-  ensure
-    puts "killing elm reactor port=#{ELM_REACTOR_PORT}; pid=#{elm_server_pid}"
-    Process.kill('KILL', elm_server_pid)
-  end
-
-  def run_sinatra(example_name)
-    port = 8901  # hardcoded in the Elm client
-
-    puts "running sinatra (#{example_name}) on port #{port}"
-    sinatra_src = "spec/zoo/#{example_name}.rb"
-
-    throw "bad sinatra example_name; could not find: #{sinatra_src}" unless File.file? sinatra_src
-
-    tmp_out = Tempfile.new("rspec-sinatra-out")
-    tmp_err = Tempfile.new("rspec-sinatra-err")
-    server_pid = Process.spawn(
-      "ruby #{sinatra_src} -p #{port}",
-      out: tmp_out.path,
-      err: tmp_err.path
-    )
-
-    yield
-  ensure
-    puts "killing sinatra (#{example_name}) port=#{port}; pid=#{server_pid}"
-    Process.kill('KILL', server_pid) if server_pid
-  end
+  include_context "logger"
+  include_context "runs elm reactor"
+  include_context "sinatra examples"
 
   # SEE: /spec/zoo/<example>.rb
   context "when PATCH /email/<from>/<timestamp>/ returns successfully" do
