+++ corpus-client/spec/spec_helper.rb
@@ -17,6 +17,8 @@
 require 'capybara'
 require 'capybara/rspec'
 
+require 'logger'
+
 require 'pry'
 
 require 'rspec'
@@ -29,6 +31,59 @@ Capybara.default_driver = :selenium
 ELM_REACTOR_PORT=8900
 
 CLIENT_PATH='/src/Main.elm'
+
+RSpec.shared_context "logger" do
+  logfile = Tempfile.new("execspec-log")
+  logger = Logger.new(logfile)
+  logger.level = Logger::INFO
+  let(:logger) { logger }
+end
+
+RSpec.shared_context "runs elm reactor" do
+  # Run/kill the elm-reactor
+  around(:example) do |example|
+    logger.info("running elm reactor on port #{ELM_REACTOR_PORT}")
+    tmp_elm_out = Tempfile.new("rspec-elm-out")
+    tmp_elm_err = Tempfile.new("rspec-elm-err")
+    elm_server_pid = Process.spawn(
+      "elm reactor --port=#{ELM_REACTOR_PORT}",
+      out: tmp_elm_out.path,
+      err: tmp_elm_err.path
+    )
+
+    Capybara.app_host = "http://localhost:#{ELM_REACTOR_PORT}"
+
+    example.run
+  ensure
+    logger.info("killing elm reactor port=#{ELM_REACTOR_PORT}; pid=#{elm_server_pid}")
+    Process.kill('KILL', elm_server_pid)
+  end
+end
+
+
+RSpec.shared_context "sinatra examples" do
+  def run_sinatra(example_name)
+    port = 8901  # hardcoded in the Elm client
+
+    logger.info("running sinatra (#{example_name}) on port #{port}")
+    sinatra_src = "spec/zoo/#{example_name}.rb"
+
+    throw "bad sinatra example_name; could not find: #{sinatra_src}" unless File.file? sinatra_src
+
+    tmp_out = Tempfile.new("rspec-sinatra-out")
+    tmp_err = Tempfile.new("rspec-sinatra-err")
+    server_pid = Process.spawn(
+      "ruby #{sinatra_src} -p #{port}",
+      out: tmp_out.path,
+      err: tmp_err.path
+    )
+
+    yield
+  ensure
+    logger.info("killing sinatra (#{example_name}) port=#{port}; pid=#{server_pid}")
+    Process.kill('KILL', server_pid) if server_pid
+  end
+end
 
 RSpec.configure do |config|
   # rspec-expectations config goes here. You can use an alternate
